split blazemeter.py into pieces
